+++ lib/locationUtil/locationDatabase.dart
@@ -1,7 +1,7 @@
 ///Latest is coming from https://mpt.i906.my/code.html
 ///All location data get from https://api.azanpro.com/reference/zone/grouped based on JAKIM website
 ///Previosly, location data is fetch from JSON https://gist.github.com/fareezMaple/02648187370acb6db11e20ee143e471e also come from same API.
-import 'package:waktusolatmalaysia/utils/location/location.dart';
+import 'location.dart';
 
 class LocationDatabase {
   static List<Location> _locationDatabase = [
