+++ lib/locationUtil/location_coordinate_model.dart
