+++ lib/main.dart
@@ -8,12 +8,11 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:provider/provider.dart';
 import 'package:timezone/data/latest.dart' as tz;
 import 'package:timezone/timezone.dart' as tz;
-import 'package:waktusolatmalaysia/utils/location/location_provider.dart';
 import 'package:waktusolatmalaysia/utils/sharing_fab.dart';
 import 'package:waktusolatmalaysia/views/onboarding_page.dart';
 import 'CONSTANTS.dart';
-import 'utils/LocationData.dart';
-import 'utils/notifications_helper.dart';
+import 'location/location_provider.dart';
+import 'notificationUtil/notifications_helper.dart';
 import 'views/Settings%20part/ThemeController.dart';
 import 'views/Settings%20part/settingsProvider.dart';
 import 'views/appBody.dart';
