--- lib/networking/ApiProvider.dart
@@ -1,39 +1,0 @@
-import 'dart:convert';
-import 'dart:async';
-import 'dart:io';
-import 'CustomException.dart';
-import 'package:http/http.dart' as http;
-
-class ApiProvider {
-  Future<dynamic> get(dynamic url) async {
-    var responseJson;
-    try {
-      final response = await http.get(url);
-      responseJson = _response(response);
-    } on SocketException {
-      throw FetchDataException('No Internet connection');
-    }
-    print('Url called: $url');
-    return responseJson;
-  }
-
-  dynamic _response(http.Response response) {
-    switch (response.statusCode) {
-      case 200:
-        var responseJson = json.decode(response.body.toString());
-        print('responseJson is $responseJson');
-        return responseJson;
-      case 400:
-        throw BadRequestException(response.body.toString());
-      case 401:
-
-      case 403:
-        throw UnauthorisedException(response.body.toString());
-      case 500:
-
-      default:
-        throw FetchDataException(
-            'Error occured while communication with server with StatusCode : ${response.statusCode}');
-    }
-  }
-}
