+++ lib/views/GetPrayerTime.dart
@@ -5,10 +5,10 @@ import 'package:flutter_spinkit/flutter_spinkit.dart';
 import 'package:fluttertoast/fluttertoast.dart';
 import 'package:get_storage/get_storage.dart';
 import 'package:provider/provider.dart';
+import 'package:waktusolatmalaysia/location/location_provider.dart';
 import 'package:waktusolatmalaysia/models/mpti906PrayerData.dart';
-import 'package:waktusolatmalaysia/networking/mpt_fetch_api.dart';
+import 'package:waktusolatmalaysia/utils/mpt_fetch_api.dart';
 import 'package:waktusolatmalaysia/utils/isolate_handler_notification.dart';
-import 'package:waktusolatmalaysia/utils/location/location_provider.dart';
 import '../CONSTANTS.dart';
 import '../utils/DateAndTime.dart';
 import '../utils/RawPrayDataHandler.dart';
