+++ lib/views/Settings part/ThemeController.dart
@@ -4,7 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
-import 'package:waktusolatmalaysia/utils/prevent_update_notifs.dart';
+import 'package:waktusolatmalaysia/notificationUtil/prevent_update_notifs.dart';
 
 class ThemeController extends GetxController {
   static ThemeController get to => Get.find();
