+++ lib/views/appBody.dart
@@ -5,8 +5,8 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:hijri/hijri_calendar.dart';
 import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
-import 'package:waktusolatmalaysia/utils/location/locationDatabase.dart';
-import 'package:waktusolatmalaysia/utils/location/location_provider.dart';
+import 'package:waktusolatmalaysia/location/location_provider.dart';
+import 'package:waktusolatmalaysia/locationUtil/locationDatabase.dart';
 import 'package:waktusolatmalaysia/views/Settings%20part/settingsProvider.dart';
 import '../utils/sizeconfig.dart';
 import 'GetPrayerTime.dart';
